Control.Monad.NState

Description

A nondeterministic state monad.

Documentation

type NState s = StateT s []

A non-deterministic state monad parameterized by the type s of the state to carry.

get :: NState s s

Return the state from the internals of the monad.

put :: s -> NState s ()

Replace the state inside the monad.

state :: (s -> (a, s)) -> NState s a

Embed a simple state action into the monad.

nstate :: (s -> [(a, s)]) -> NState s a

Embed a non-deterministic function state function, realized via the type s -> [(a,s)], into the NState monad.

modify :: (s -> s) -> NState s ()

Maps an old state to a new state inside a state monad. The old state is thrown away.

branch

Arguments

:: Foldable t

=> t a
list of alternatives

-> NState s a
nondeterministic state computation

Non-deterministically select a value from the collection of alternatives.

runNState

Arguments

:: NState s a
nondeterministic state computation

-> s
initial state

-> [(a, s)]
list of return values and final states

Unwrap an NState computation as a function.

evalNState

Arguments

:: NState s a
nondeterministic state computation

-> s
initial state

-> [a]
list of return values

Evaluate a non-deterministic state computation with the given initial state and return the list of final values, discarding the final states.

execNState

Arguments

:: NState s a
nondeterministic state computation

-> s
initial state

-> [s]
list of return states

Evaluate a non-deterministic state computation with the given initial state and return the list of final states, discarding the final vailes.